Number of Islands
about this problem where we have a number of islands where given in matrix some of them may be connected. This question is to find the islands surrounded by water. So here we will be checking all 8 direction of the island with the island method. If there exits an island in atleast one direction it it forms a connection which creates a network of island. Our goal is to find the number of island which are surrounded by only water.
 NOTE:grp of islands is considered as one island